improve deployment detail layout
